feat(subwindow.py): move addstr() and cs_addstr() to window.py
